requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
cryptg==0.2.post4
discord.py==2.0.0a4275+g544bb1e2
flanautils
frozenlist==1.2.0
hachoir==3.1.2
idna==3.3
iso8601==1.0.2
jellyfish==0.9.0
kaleido==0.2.1
mpmath==1.2.1
multidict==5.2.0
Pillow==9.0.0
plotly==5.5.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pymongo==4.0.1
rsa==4.8
six==1.16.0
sympy==1.9
Telethon==1.24.0
tenacity==8.0.1
twitchio==2.1.4
typing-extensions==4.0.1
yarl==1.7.2
